TravelGo/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
boto3==1.34.131
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
gunicorn==22.0.0
botocore==1.34.131
